extract messages into message component
