+++ client/src/views/DiscussionView.tsx
@@ -4,6 +4,7 @@ import { deleteDiscussion, getChatroomRef } from "../res/BackendConnection";
 import firebase from 'firebase/app';
 import { message, userData } from "../res/interfaces";
 import NotFound from "./NotFound";
+import Message from "../Message";
 
 const DiscussionView = (props: {userData: userData}) => {
 
@@ -45,23 +46,9 @@ const DiscussionView = (props: {userData: userData}) => {
       alert("Please set a username in your settings before sending a message.");
       return;
     }
-    chatroomRef.collection("Messages").add({ content: messageText, time: firebase.firestore.Timestamp.now(), author: userData.username });
+    chatroomRef.collection("Messages").add({ content: messageText, time: firebase.firestore.Timestamp.now(), author: userData.PUID });
     setMessageText("");
   };
-
-  const pad =(num: number, size: number) => {
-    let str = num.toString();
-    while (str.length < size) str = "0" + num;
-    return str;
-}
-
-  const formatDate = (date: Date) => {
-    const now = Date.now();
-    if(now - date.getTime() < 86400000){
-      return `${date.getHours()}:${pad(date.getMinutes(), 2)}`;
-    }
-    return `${date.getMonth()}/${pad(date.getDate(), 2)}/${date.getFullYear()}`;
-  }
 
   if(!validDiscussion){
     return(
@@ -79,18 +66,12 @@ const DiscussionView = (props: {userData: userData}) => {
       
       {messageList?.map((message, i) => {
         return (
-          <div className= "msgClass" key={i}>
-            <p>
-              <span className="msg-text">
-                {`${message.author}: ${message.content}`}
-              </span>
-              <span className="msg-time">
-                {formatDate(message.time)}
-              </span>
-            </p>
+          <div key={i}>
+            <Message message={message}/>
           </div>
         );
       })}
+      
       <form>
         <input className="send-msg-form"
           type="text"
